+++ DataConnectors/Digital shadows/.funcignore
@@ -1,0 +1,5 @@
+.git*
+.vscode
+local.settings.json
+test
+.venv
