+++ DataConnectors/Digital shadows/DigitalShadowsConnectorAzureFunction/AS_poller.py
@@ -1,0 +1,4 @@
+""" polls data from azure sentinel incidents to DS """
+
+class poller:
+    pass
