+++ DataConnectors/Digital shadows/DigitalShadowsConnectorAzureFunction/function.json
@@ -1,0 +1,12 @@
+{
+  "scriptFile": "__init__.py",
+  "bindings": [
+    {
+      "name": "mytimer",
+      "type": "timerTrigger",
+      "runOnStartup": true,
+      "direction": "in",
+      "schedule": "0 */5 * * * *"
+    }
+  ]
+}
