--- DataConnectors/digital shadows/.funcignore
@@ -1,5 +1,0 @@
-.git*
-.vscode
-local.settings.json
-test
-.venv
