--- DataConnectors/digital shadows/proxies.json
@@ -1,4 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/proxies",
-  "proxies": {}
-}
