--- DataConnectors/digital shadows/requirements.txt
@@ -1,7 +1,0 @@
-# DO NOT include azure-functions-worker in this file
-# The Python Worker is managed by Azure Functions platform
-# Manually managing azure-functions-worker may cause unexpected issues
-
-azure-functions
-azure-storage-file-share
-requests
